Transformers extracted.
