+++ src/components/portfolio/PortfolioDialog.vue
@@ -80,7 +80,7 @@
 import { defineComponent, PropType, computed, ref, toRef, Ref } from 'vue';
 import { useLoadingStore } from 'stores/loading';
 import { usePortfolioStore } from 'src/stores/portfolios';
-import portfolioAPI, { viewTransformer } from 'src/service/portfolio';
+import portfolioAPI from 'src/service/portfolio';
 import { Portfolio } from 'app/shared/types';
 
 const emptyPortfolioTemplate = (): Portfolio => ({
@@ -174,7 +174,6 @@ export default defineComponent({
       isNew,
       initialDeposit,
       localPortfolio,
-      viewTransformer,
       setLocalPortfolio,
       submitForm,
     };
