+++ src/service/holdings.ts
@@ -1,63 +1,12 @@
-import { Quote, Holding, HoldingWithProfits, Transaction } from 'shared/types';
+import { Holding, Transaction } from 'shared/types';
+import { transactionsTransformer } from 'app/shared/transformers';
 import {
   getCollections,
   firestoreAPI,
   queries,
 } from 'src/service/firebase/collections';
-import { transformer as transactionTransformer } from 'src/service/transactions';
 
 const holdingsCollection = () => getCollections().holding;
-
-export const transformer = {
-  currentValue: (holding: Holding, quote: Quote) =>
-    holding.shares * quote.regularMarketPrice,
-  totalValue: (holding: Holding) => holding.shares * holding.avgPrice,
-  profit: (holding: Holding, quote: Quote) => {
-    const profitValue =
-      transformer.currentValue(holding, quote) -
-      transformer.totalValue(holding) +
-      (holding?.realizedProfits ?? 0);
-
-    return {
-      value: profitValue,
-      percent: Math.abs(profitValue / holding.invested),
-    };
-  },
-  dailyChange: (holding: Holding, quote: Quote) => {
-    const dailyChangeValue = quote.regularMarketChange * holding.shares;
-
-    return {
-      value: dailyChangeValue,
-      percent: quote.regularMarketChangePercent / 100,
-    };
-  },
-  summary: (holdings: HoldingWithProfits[]) =>
-    holdings.reduce(
-      (acc, holding) => {
-        acc.shares += holding.shares;
-        acc.profit += holding.profit.value;
-        acc.currentValue += holding.currentValue;
-        acc.invested += holding.invested;
-        acc.realized += holding.realizedProfits ?? 0;
-        acc.captialGains +=
-          holding.profit.value - (holding.realizedProfits ?? 0);
-        acc.dailyChange += holding.dailyChange.value;
-        acc.fees += holding?.fees ?? 0;
-
-        return acc;
-      },
-      {
-        shares: 0,
-        profit: 0,
-        currentValue: 0,
-        invested: 0,
-        realized: 0,
-        captialGains: 0,
-        dailyChange: 0,
-        fees: 0,
-      }
-    ),
-};
 
 const api = {
   list: async (portfolioId: string) =>
@@ -99,7 +48,7 @@ const api = {
     );
 
     const buyTransactions = holdingTransactions.filter(
-      transactionTransformer.isBuy
+      transactionsTransformer.isBuy
     );
 
     // Total and Average price are calculated by current Buy transactions
@@ -111,14 +60,14 @@ const api = {
     // Fees included
     holding.avgPrice = totalShares
       ? (buyTransactions.reduce(
-          (acc, t) => acc + transactionTransformer.actualValue(t),
+          (acc, t) => acc + transactionsTransformer.actualValue(t),
           0
         ) ?? 0) / totalShares
       : 0;
 
     // Calculate invested by transactions funds (original shares), Fees included
     holding.invested = buyTransactions.reduce(
-      (acc, t) => acc + transactionTransformer.totalValue(t),
+      (acc, t) => acc + transactionsTransformer.totalValue(t),
       0
     );
