+++ src/stores/portfolios.ts
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia';
+import { holdingsTransformer } from 'app/shared/transformers';
 import portfolioAPI from 'src/service/portfolio';
-import { transformer as holdingsTransformer } from 'src/service/holdings';
 import { Deposit, Portfolio } from 'app/shared/types';
 import { useTransactionsStore } from 'stores/transactions';
 import { useHoldingsStore } from 'stores/holdings';
