+++ src/stores/transactions.ts
@@ -1,5 +1,6 @@
 import { defineStore } from 'pinia';
-import transactionsAPI, { transformer } from 'src/service/transactions';
+import { transactionsTransformer as transformer } from 'app/shared/transformers';
+import transactionsAPI from 'src/service/transactions';
 import { Transaction } from 'app/shared/types';
 import { useQuotesStore } from 'stores/quotes';
